+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Fix paint glitches of order list and pattern editor
 - Fix edit step crashing on change
 - [#190] - Fix the error of FM envelope/operator paste when "Show FM detune as signed" enabled (thanks [@nyanpasu64])
+- [#191] - Enable copy and paste in the order list without pattern edit mode (thanks [@nyanpasu64])
 
 [@freq-mod]: https://github.com/freq-mod
 [@RigidatoMS]: https://twitter.com/RigidatoMS
@@ -34,6 +35,7 @@
 [#187]: https://github.com/rerrahkr/BambooTracker/issues/187
 [#189]: https://github.com/rerrahkr/BambooTracker/issues/189
 [#190]: https://github.com/rerrahkr/BambooTracker/issues/190
+[#191]: https://github.com/rerrahkr/BambooTracker/issues/191
 
 ## v0.3.5 (2020-02-15)
 ### Added